server/ui updates
A handful of modifications.  (1) Changes to UI layout -- this still
needs work.  (2) Changes to plotting so that sampled points will show up
on top of available points.  (3) Changes to server code forcing order of
operations (i.e. do NOT draw sample until run button is pressed).  (4)
Slight modifications to file export.  Please not that file export DOES
NOT work through RStudio interface - you need to run the app in the
browser to use this functionality.  (5) shinydashboard package is used
here so we will need to add to the DESCRIPTION file if we want to keep
using it.